+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+next release (1.5.0)
+--------------------
+
+* Add consumer registry based on Psr Container.
+* Change constructor signature for `RoundRobinConsumer` (pass registry instead of list of consumers). 
+
 v1.4.2
 ------